requirements.txt
flask
requests
groq
python-dotenv
gunicorn